Move non-public things out of api.py
Move h5py wrapper classes into a new wrappers.py and spaceid_to_slice() into
slicetools.py.
