+++ versioned_hdf5/tests/test_api.py
@@ -3,10 +3,8 @@ from pytest import raises
 import numpy as np
 from numpy.testing import assert_equal
 
-from h5py._hl.selections import Selection
-
 from ..backend import DEFAULT_CHUNK_SIZE
-from ..api import VersionedHDF5File, spaceid_to_slice
+from ..api import VersionedHDF5File
 
 from .test_backend import setup
 
@@ -328,40 +326,6 @@ def test_delete():
         assert set(file['version2']) == {'test_data'}
         assert_equal(file['version2']['test_data'], test_data)
         assert file['version2'].datasets().keys() == {'test_data'}
-
-def test_spaceid_to_slice():
-    with setup() as f:
-        shape = 10
-        a = f.create_dataset('a', data=np.arange(shape))
-
-        for start in range(0, shape):
-            for count in range(0, shape):
-                for stride in range(1, shape):
-                    for block in range(0, shape):
-                        if count != 1 and block != 1:
-                            # Not yet supported. Doesn't seem to be supported
-                            # by HDF5 either (?)
-                            continue
-
-                        spaceid = a.id.get_space()
-                        spaceid.select_hyperslab((start,), (count,),
-                                                 (stride,), (block,))
-                        sel = Selection((shape,), spaceid)
-                        try:
-                            a[sel]
-                        except ValueError:
-                            # HDF5 doesn't allow stride/count combinations
-                            # that are impossible (the count must be the exact
-                            # number of elements in the selected block).
-                            # Rather than trying to enumerate those here, we
-                            # just check what doesn't give an error.
-                            continue
-                        try:
-                            s = spaceid_to_slice(spaceid)
-                        except:
-                            print(start, count, stride, block)
-                            raise
-                        assert_equal(a[s], a[sel], f"{(start, count, stride, block)}")
 
 def test_resize():
     with setup() as f:
